% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indi_search.R
\name{indi_search}
\alias{indi_search}
\title{List Filenames Containing a Specific Search Word}
\usage{
indi_search(searchWord)
}
\arguments{
\item{searchWord}{A character string representing the pattern to search for in file names.}
}
\value{
A character vector of file names containing the search pattern. If no files are found a message appears in the console.
}
\description{
This function searches for and returns indicator names in the data directory set by the option `indicatorDataPath` that contain a specific search word.
}
\examples{
# List all files in the directory containing 'biodiversity' in the name
searchResults <- indi_search("biodiversity")
print(searchResults)

}
