% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indi_summary.R
\name{indi_summary}
\alias{indi_summary}
\title{Load Comprehensive Summary of Indicator Metadata}
\usage{
indi_summary()
}
\value{
A data frame containing comprehensive metadata for the available indicators of the package.
}
\description{
This function loads a detailed overview of the metadata for indicators available in the package. The summary includes key information such as the indicator name, data sources, time span covered, EBV class, and other relevant attributes.
}
\examples{
# Load the comprehensive metadata for the indicators
Tab1.dat <- indi_summary()
head(Tab1.dat)

}
\references{
IPBES (2019). *Global Assessment Chapter 2.2 – Supplementary Material: Indicators of Status & Trends in Nature*.
}
