% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_indicators_by_searchword.R
\name{list_indicators_by_searchword}
\alias{list_indicators_by_searchword}
\title{List Filenames Containing a Specific Search Word}
\usage{
list_indicators_by_searchword(searchWord)
}
\arguments{
\item{searchWord}{A character string representing the word to search for in filenames.}
}
\value{
A character vector of filenames containing the search word, or a message if no files are found.
}
\description{
This function searches for and returns filenames in the data directory set by the option `indicatorDataPath` that contain a specific search word.
}
\examples{
# List all files in the directory containing 'biodiversity' in the name
searchResults <- list_indicators_by_searchword("biodiversity")
print(searchResults)

}
