% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_indicatorData.R
\name{load_indicatorData}
\alias{load_indicatorData}
\title{Load Indicator Data}
\usage{
load_indicatorData(indicatorName = "percentage_of_live_coral_cover")
}
\arguments{
\item{indicatorName}{A character string representing the name of the indicator to be loaded. To view all available indicators, use the `print_indicatorNames()` function.}
}
\value{
A data frame containing the year and the corresponding global indicator values.
}
\description{
This function loads indicator data either from the package’s included data or from a local directory as specified by the `options(indicatorDataPath = /inst/data)` setting (default location). The data provides global averages over time for the specified indicator.
}
\examples{
# Load data for a specific indicator
dat <- load_indicatorData("Aboveground_biomass_PgC.csv")
head(dat)
plot(dat)

}
\references{
Data sourced from IPBES (2019). *Global Assessment Chapter 2.2 – Supplementary Material: Indicators of Status & Trends in Nature*.
}
