% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indi_data.R
\name{plot_indi_data}
\alias{plot_indi_data}
\title{Plot raw Values of Indicator values from file}
\usage{
plot_indi_data(indicatorName)
}
\arguments{
\item{indicatorName}{A character string representing the full name of the indicator to be plotted. To view a list of available indicators, use the `indi_search()` function.}
}
\value{
A ggplot figure with "Year" on the x-axis and "Raw value" on the y-axis, representing the absolute values of the selected indicator.
}
\description{
This function generates a plot showing the absolute values of a given indicator over time, with years on the x-axis and raw indicator values on the y-axis.
}
\examples{
# Example: Plot the absolute values of the Aboveground biomass indicator
plot_indi_data("Aboveground_biomass.txt")

}
