% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scaled_to_remaining.R
\name{plot_scaled_to_remaining}
\alias{plot_scaled_to_remaining}
\title{Plot Percentage Remaining of the Indicator Over Time}
\usage{
plot_scaled_to_remaining(indicatorName)
}
\arguments{
\item{indicatorName}{A character string representing the full name of the indicator to be plotted. Use `print_indicatorNames()` to view a list of available indicators.}
}
\value{
A ggplot object displaying the trend line with the x-axis representing "Year" and the y-axis representing either the raw or rescaled values.
}
\description{
This function generates a trend line plot for a given indicator over time. If applicable, the data is rescaled to a percentage scale where 100 represents the "pristine" state and 0 represents complete depletion. For details on the calculation see the function `rescale_to_percent()`
}
\examples{
# Example: Plot the trend line for the Percentage of Natural Habitat Extent
plot_scaled_to_remaining("aboveground_biomass")
}
